Config validation and settings map (#23)
* Update demo config and html styles, add serve dependency

* Regen

* Split off defaults, add settings, spacing keys

* rm

* nope

* Update config

* Add config validation test

* Add expected version test

* Generate settings

+++ src/index.js
@@ -8,6 +8,10 @@ const { transform } = require('./transform-config')
 const { merge } = require('./merge-objects')
 const { write: writeMixins } = require('./write-mixins')
 const { write: writeStyles } = require('./write-styles')
+const { validate } = require('./validate-config')
+const { get: getSassSettingsMap } = require('./get-sass-settings-map')
+
+const { CURRENT_FORMAT_VERSION = 1 } = process.env
 
 async function main({ config, output, compression } = {}) {
   // messy entry paths — TODO: tidy/structure
@@ -15,6 +19,10 @@ async function main({ config, output, compression } = {}) {
   const tmpDir = path.join(__dirname, '../build')
   const defaultsPath = path.join(__dirname, './defaults.yml')
   const tmpDirSass = path.join(tmpDir, 'sass')
+  const generatedSassSettingsPath = path.join(
+    tmpDirSass,
+    'core/_generated-settings.scss'
+  )
   const generatedSassMixinsPath = path.join(
     tmpDirSass,
     'core/_generated-mixins.scss'
@@ -29,13 +37,17 @@ async function main({ config, output, compression } = {}) {
   const parsed = await parse(config)
   // parse the deafults (internal shadow config file)
   const defaults = await parse(defaultsPath)
-  // TODO: assert validation of config file before attempting merge
-  // validateConfigorThrow(parsed)
-  // since config is valid, clean intermediate target for built files
-  await fs.remove(tmpDir)
   // merge custom config onto necessary but optional defaults (e.g.
   // prefixes and vertical rhythm settings)
   const merged = merge(defaults, parsed)
+  // Validate the combined config to ensure any assumed values are
+  // set
+  const [valid, { message }] = validate(merged, CURRENT_FORMAT_VERSION)
+  if (!valid) throw new Error(message)
+  // since config is valid, clean intermediate target for built files
+  await fs.remove(tmpDir)
+  // convert config settings to a sass map
+  const settings = getSassSettingsMap(merged)
   // fill in blanks for styles (i.e. use null instead of undefined for
   // omitted mixin properties)
   const sanitized = await sanitize(merged)
@@ -47,6 +59,8 @@ async function main({ config, output, compression } = {}) {
   // copy the source sass dir to its temp destination before we start
   // to mutate it with custom generated code
   await fs.copy(srcDir, tmpDirSass)
+  // write the generated settings to the temp sass dir
+  await fs.writeFile(generatedSassSettingsPath, settings)
   // write the generated mixins to the temp sass dir
   await fs.writeFile(generatedSassMixinsPath, mixins)
   // write the classes that refer to the generated mixins (as string)
